MK1.9: engine measures masks in display-corrected source pixels
Export mask stacks, add_mask and mask motion evidence read
AssetMedia.display_size (PAR and rotation applied), the same space editor-core
draws in. A left-half mask on a rotated or anamorphic asset cuts the left half
of the frame MoviePy decodes; square unrotated media is unchanged (legacy and
golden mask render tests pass).

+++ engine/python/framepilot_engine/render/masks.py
@@ -186,7 +186,7 @@ def mask_scalar_at(mask: Any, property_name: str, source_time: float) -> float |
 
 
 def mask_frame_box(
-    mask: Any, media_size: tuple[int, int] | None, source_time: float
+    mask: Any, media_size: tuple[float, float] | None, source_time: float
 ) -> tuple[float, float, float, float] | None:
     """A rectangle/ellipse mask's ``(x, y, width, height)`` as fractions of the source picture.
